requirements.txt
django
djangorestframework
django-crispy-forms
django_forms_bootstrap
django-bootstrap4
python-dotenv